+++ modules/crowdfundings/__tests__/integration.test.js
@@ -1,24 +1,6 @@
 require('dotenv').config()
 const test = require('tape-async')
-const { apolloFetch, connectIfNeeded, pgDatabase } = require('./helpers.js')
-
-const SUBMIT_PLEDGE_USER = {
-  'firstName': 'pascal',
-  'lastName': 'kaufmann',
-  'email': 'pascal@reactive.one'
-}
-
-const PLEDGE_MUTATION = `
-  mutation submitPledge($total: Int!, $options: [PackageOptionInput!]!, $user: UserInput!, $reason: String) {
-    submitPledge(pledge: {total: $total, options: $options, user: $user, reason: $reason}) {
-      pledgeId
-      userId
-      emailVerify
-      pfAliasId
-      pfSHA
-    }
-  }
-`
+const { apolloFetch, connectIfNeeded } = require('./helpers.js')
 
 test('setup', async (t) => {
   await connectIfNeeded()
@@ -160,255 +142,5 @@ test('test data exists', async (t) => {
       ]
     }
   }, 'test data exists and looks good')
-  t.end()
-})
-
-test('default pledge with ABO package', async (t) => {
-  await connectIfNeeded()
-  pgDatabase().public.pledges.truncate({ cascade: true })
-
-  const result = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 24000,
-      'options': [{
-        'amount': 1,
-        'price': 24000,
-        'templateId': '00000000-0000-0000-0008-000000000001'
-      }],
-      'user': SUBMIT_PLEDGE_USER
-    }
-  })
-  t.notOk(result.errors, 'graphql query successful')
-  t.notEqual(result.data.submitPledge.pledgeId, null, 'pledgeId returned')
-  t.notEqual(result.data.submitPledge.userId, null, 'userId returned')
-  t.notEqual(result.data.submitPledge.pfAliasId, null, 'alias id returned')
-  t.equal(result.data.submitPledge.emailVerify, null, 'email must not be verified, because user does not exist')
-  t.ok(result.data.submitPledge.pfSHA, 'hash returned')
-
-  const pledge = await pgDatabase().public.pledges.findOne({ id: result.data.submitPledge.pledgeId })
-  t.equal(pledge.total, 24000, 'correct total value in db')
-
-  t.comment('submit pledge again with the same e-mail address')
-  const result2 = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 24000,
-      'options': [{
-        'amount': 1,
-        'price': 24000,
-        'templateId': '00000000-0000-0000-0008-000000000001'
-      }],
-      'user': SUBMIT_PLEDGE_USER
-    }
-  })
-  t.notOk(result2.errors, 'graphql query successful')
-  t.equal(result2.data.submitPledge.pledgeId, null, 'no pledgeId returned')
-  t.equal(result2.data.submitPledge.userId, null, 'userId is null')
-  t.equal(result2.data.submitPledge.pfAliasId, null, 'aliasId is null')
-  t.equal(result2.data.submitPledge.emailVerify, true, 'email must be verified, because e-mail')
-  t.equal(result2.data.submitPledge.pfSHA, null, 'no hash')
-  t.end()
-})
-
-test('pledge with ABO package and donation', async (t) => {
-  await connectIfNeeded()
-  pgDatabase().public.pledges.truncate({ cascade: true })
-
-  const result = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 40000,
-      'options': [{
-        'amount': 1,
-        'price': 24000,
-        'templateId': '00000000-0000-0000-0008-000000000001'
-      }],
-      'user': SUBMIT_PLEDGE_USER
-    }
-  })
-  t.notOk(result.errors, 'graphql query successful')
-  t.notEqual(result.data.submitPledge.pledgeId, null, 'pledgeId returned')
-  t.ok(result.data.submitPledge.pfSHA, 'hash returned')
-
-  const pledge = await pgDatabase().public.pledges.findOne({ id: result.data.submitPledge.pledgeId })
-  t.equal(pledge.donation, 16000, 'correct donation value in db')
-  t.equal(pledge.total, 40000, 'correct total value in db')
-  t.end()
-})
-
-test('pledge with ABO package and userPrice', async (t) => {
-  await connectIfNeeded()
-  pgDatabase().public.pledges.truncate({ cascade: true })
-
-  const result = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 10000,
-      'options': [{
-        'amount': 1,
-        'price': 24000,
-        'templateId': '00000000-0000-0000-0008-000000000001'
-      }],
-      'user': SUBMIT_PLEDGE_USER,
-      'reason': 'student'
-    }
-  })
-  t.notOk(result.errors, 'graphql query successful')
-  t.notEqual(result.data.submitPledge.pledgeId, null, 'pledgeId returned')
-  t.ok(result.data.submitPledge.pfSHA, 'hash returned')
-
-  const pledge = await pgDatabase().public.pledges.findOne({ id: result.data.submitPledge.pledgeId })
-  t.equal(pledge.donation, -14000, 'minus donation value in db')
-  t.equal(pledge.total, 10000, 'correct total value in db')
-  t.end()
-})
-
-test('pledge with userPrice but no reason', async (t) => {
-  await connectIfNeeded()
-  pgDatabase().public.pledges.truncate({ cascade: true })
-  const result = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 10000,
-      'options': [{
-        'amount': 1,
-        'price': 24000,
-        'templateId': '00000000-0000-0000-0008-000000000001'
-      }],
-      'user': SUBMIT_PLEDGE_USER
-    }
-  })
-  t.ok(result.errors, 'no reason, no pledge')
-  t.end()
-})
-
-test('pledge with ABO package and userPrice too low (minUserPrice = 1000) is not possible', async (t) => {
-  await connectIfNeeded()
-  pgDatabase().public.pledges.truncate({ cascade: true })
-
-  const result = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 999,
-      'options': [{
-        'amount': 1,
-        'price': 999,
-        'templateId': '00000000-0000-0000-0008-000000000001'
-      }],
-      'user': SUBMIT_PLEDGE_USER,
-      'reason': 'student'
-    }
-  })
-  t.ok(result.errors, 'userPrice too low, no pledge')
-  t.end()
-})
-
-test('pledge with 2 x ABO (maxAmount = 1) is not possible', async (t) => {
-  await connectIfNeeded()
-  pgDatabase().public.pledges.truncate({ cascade: true })
-
-  const result = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 40000,
-      'options': [{
-        'amount': 2,
-        'price': 24000,
-        'templateId': '00000000-0000-0000-0008-000000000001'
-      }],
-      'user': SUBMIT_PLEDGE_USER,
-      'reason': 'student'
-    }
-  })
-  t.ok(result.errors, 'amount too high, no pledge')
-  t.end()
-})
-
-test('pledge with PATRON and 1 x SWEETS (minAmount = 2) is not possible', async (t) => {
-  await connectIfNeeded()
-  pgDatabase().public.pledges.truncate({ cascade: true })
-
-  const result = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 103000,
-      'options': [{
-        'amount': 1,
-        'price': 100000,
-        'templateId': '00000000-0000-0000-0008-000000000003'
-      }, {
-        'amount': 1,
-        'price': 3000,
-        'templateId': '00000000-0000-0000-0008-000000000004'
-      }],
-      'user': SUBMIT_PLEDGE_USER
-    }
-  })
-  t.ok(result.errors, 'amount too low, no pledge')
-  t.end()
-})
-
-test('pledge with PATRON and no SWEETS (minAmount = 2) is not possible', async (t) => {
-  await connectIfNeeded()
-  pgDatabase().public.pledges.truncate({ cascade: true })
-
-  const result = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 103000,
-      'options': [{
-        'amount': 1,
-        'price': 100000,
-        'templateId': '00000000-0000-0000-0008-000000000003'
-      }],
-      'user': SUBMIT_PLEDGE_USER
-    }
-  })
-  t.ok(result.errors, 'mandatory package option does not exist, no pledge')
-  t.end()
-})
-
-test('pledge with PATRON package (userPrice = false) and a total that is lower than the price is not possible', async (t) => {
-  await connectIfNeeded()
-  pgDatabase().public.pledges.truncate({ cascade: true })
-
-  const result = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 99999,
-      'options': [{
-        'amount': 1,
-        'price': 10000,
-        'templateId': '00000000-0000-0000-0008-000000000003'
-      }],
-      'user': SUBMIT_PLEDGE_USER
-    }
-  })
-  t.ok(result.errors, 'total lower than price, no pledge')
-  t.end()
-})
-
-test('pledge with mixed PATRON package options and ABO package option is not possible', async (t) => {
-  await connectIfNeeded()
-  pgDatabase().public.pledges.truncate({ cascade: true })
-
-  const result = await apolloFetch({
-    query: PLEDGE_MUTATION,
-    variables: {
-      'total': 100000,
-      'options': [{
-        'amount': 1,
-        'price': 100000,
-        'templateId': '00000000-0000-0000-0008-000000000002'
-      }, {
-        'amount': 1,
-        'price': 20000,
-        'templateId': '00000000-0000-0000-0008-000000000001'
-      }],
-      'user': SUBMIT_PLEDGE_USER
-    }
-  })
-  t.ok(result.errors, 'cannot mix options of two packages, no pledge')
   t.end()
 })
